backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
pytesseract>=0.3.10
scrapy>=2.11.0
selenium>=4.15.0
scrapy-selenium>=0.0.7
pymongo>=4.6.0
pdfplumber>=0.10.0
fpdf>=1.7.2
